scripts/researcher/pyproject.toml
[project]
name = "playground-researcher"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.60.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
researcher = "researcher:main"
